requirements.txt
click>=8.1
pyyaml>=6.0
pytest>=7.0
